+++ docusign/src/types/types.ts
@@ -1,0 +1,13 @@
+export interface DocumentMetadata {
+  cid: string;
+  size?: number;
+  filename?: string;
+  uploadDate?: string;
+  contentType?: string;
+  links?: Array<{ Name: string; Size: number; Hash: string }>;
+}
+
+export interface PDFViewerProps {
+  fileUrl?: string;
+  height?: string;
+}
